Diverse Optimierungen, insb. fuer shape_function, die jetzt 2-3 mal so schnell sein sollte. Achtung: shape_function radius ist jetzt ein INTEGER >=2. Außerdem: Komplette Partikeleinsetzausgabe jetzt ueber IterDisplayStep gesteuert, inkl. allreduce und Fehlerausgabe.
git-svn-id: svn+ssh://iagpc151.iag.uni-stuttgart.de/usr/local/svnroot/Boltzplatz/trunk@192 f8cfc4a6-4e3a-43c3-8588-810d72ca601e

+++ src/particles/pic/deposition/pic_depo_vars.f90
@@ -20,7 +20,7 @@ REAL                                  :: r_sf               ! cutoff radius of s
 REAL                                  :: r2_sf              ! cutoff radius of shape function * cutoff radius of shape function
 REAL                                  :: r2_sf_inv          ! 1/cutoff radius of shape function * cutoff radius of shape function
 REAL                                  :: w_sf               ! shapefuntion weight
-REAL                                  :: alpha_sf           ! shapefuntion exponent 
+INTEGER                               :: alpha_sf           ! shapefuntion exponent 
 REAL                                  :: BGMdeltas(3)       ! Backgroundmesh size in x,y,z
 REAL                                  :: FactorBGM(3)       ! Divider for BGM (to allow real numbers)
 REAL                                  :: BGMVolume          ! Volume of a BGM Cell
